LayoutTests/imported/w3c/web-platform-tests/2dcontext/compositing/2d.composite.operation.darker-expected.txt

2d.composite.operation.darker
Actual output:

FAIL Canvas test: 2d.composite.operation.darker assert_equals: ctx.globalCompositeOperation === 'xor' (got darker[string], expected xor[string]) expected "xor" but got "darker"
